+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # babel-preset-dlight
 
+## 1.0.0-alpha.30
+
+### Patch Changes
+
+- refactor: rebuild unmount with store
+
 ## 1.0.0-alpha.29
 
 ### Patch Changes